// Hippocamp.Web/src/app/shared/models/enums/treatment-bmp-modeling-type.enum.ts
export enum TreatmentBMPModelingType {
  BioinfiltrationBioretentionWithRaisedUnderdrain = 1,
  BioretentionWithNoUnderdrain = 2,
  BioretentionWithUnderdrainAndImperviousLiner = 3,
  CisternsForHarvestAndUse = 4,
  ConstructedWetland = 5,
  DryExtendedDetentionBasin = 6,
  DryWeatherTreatmentSystems = 7,
  Drywell = 8,
  FlowDurationControlBasin = 9,
  FlowDurationControlTank = 10,
  HydrodynamicSeparator = 11,
  InfiltrationBasin = 12,
  InfiltrationTrench = 13,
  LowFlowDiversions = 14,
  PermeablePavement = 15,
  ProprietaryBiotreatment = 16,
  ProprietaryTreatmentControl = 17,
  SandFilters = 18,
  UndergroundInfiltration = 19,
  VegetatedFilterStrip = 20,
  VegetatedSwale = 21,
  WetDetentionBasin = 22
}